+++ tests/HasRolesTest.php
@@ -34,6 +34,60 @@ class HasRolesTest extends TestCase
     /**
      * @test
      */
+    public function it_can_sync_roles_from_a_string()
+    {
+        $this->testUser->assignRole('testRole');
+
+        $this->testUser->syncRoles('testRole2');
+
+        $this->assertFalse($this->testUser->hasRole('testRole'));
+
+        $this->assertTrue($this->testUser->hasRole('testRole2'));
+    }
+
+    /**
+     * @test
+     */
+    public function it_can_sync_multiple_roles()
+    {
+        $this->testUser->syncRoles('testRole', 'testRole2');
+
+        $this->assertTrue($this->testUser->hasRole('testRole'));
+
+        $this->assertTrue($this->testUser->hasRole('testRole2'));
+    }
+
+    /**
+     * @test
+     */
+    public function it_can_sync_multiple_roles_from_an_array()
+    {
+        $this->testUser->syncRoles(['testRole', 'testRole2']);
+
+        $this->assertTrue($this->testUser->hasRole('testRole'));
+
+        $this->assertTrue($this->testUser->hasRole('testRole2'));
+    }
+
+    /**
+     * @test
+     */
+    public function it_will_remove_all_roles_when_an_empty_array_is_past_to_sync_roles()
+    {
+        $this->testUser->assignRole('testRole');
+
+        $this->testUser->assignRole('testRole2');
+
+        $this->testUser->syncRoles([]);
+
+        $this->assertFalse($this->testUser->hasRole('testRole'));
+
+        $this->assertFalse($this->testUser->hasRole('testRole2'));
+    }
+
+    /**
+     * @test
+     */
     public function it_can_assign_a_role_using_an_object()
     {
         $this->testUser->assignRole($this->testRole);
@@ -92,13 +146,13 @@ class HasRolesTest extends TestCase
 
         $this->refreshTestUser();
 
-        $this->assertFalse($this->testUser->hasAllRoles($roleModel->all()));
+        $this->assertFalse($this->testUser->hasAllRoles(['testRole', 'second role']));
 
         $this->testUser->assignRole('second role');
 
         $this->refreshTestUser();
 
-        $this->assertTrue($this->testUser->hasAllRoles($roleModel->all()));
+        $this->assertTrue($this->testUser->hasAllRoles(['testRole', 'second role']));
     }
 
     /**